[기능개선][1,2단계] TestNumberGenerator test 패키지로 이동
